requirements.txt
streamlit
python-pptx
fonttools
